+++ CesiumRasterOverlays/src/RasterOverlayTileProvider.cpp
@@ -44,10 +44,12 @@ RasterOverlayTileProvider::RasterOverlayTileProvider(
     const CesiumUtility::IntrusivePointer<const RasterOverlay>& pOwner,
     const CesiumAsync::AsyncSystem& asyncSystem,
     const std::shared_ptr<IAssetAccessor>& pAssetAccessor,
+    const std::shared_ptr<CesiumUtility::CreditSystem>& pCreditSystem,
     const CesiumGeospatial::Ellipsoid& ellipsoid) noexcept
     : _pOwner(const_intrusive_cast<RasterOverlay>(pOwner)),
       _asyncSystem(asyncSystem),
       _pAssetAccessor(pAssetAccessor),
+      _pCreditSystem(pCreditSystem),
       _credit(std::nullopt),
       _pPrepareRendererResources(nullptr),
       _pLogger(nullptr),
@@ -57,12 +59,14 @@ RasterOverlayTileProvider::RasterOverlayTileProvider(
       _pPlaceholder(new RasterOverlayTile(*this)),
       _tileDataBytes(0),
       _totalTilesCurrentlyLoading(0),
-      _throttledTilesCurrentlyLoading(0) {}
+      _throttledTilesCurrentlyLoading(0),
+      _destructionCompleteDetails() {}
 
 RasterOverlayTileProvider::RasterOverlayTileProvider(
     const CesiumUtility::IntrusivePointer<const RasterOverlay>& pOwner,
     const CesiumAsync::AsyncSystem& asyncSystem,
     const std::shared_ptr<IAssetAccessor>& pAssetAccessor,
+    const std::shared_ptr<CesiumUtility::CreditSystem>& pCreditSystem,
     std::optional<Credit> credit,
     const std::shared_ptr<IPrepareRasterOverlayRendererResources>&
         pPrepareRendererResources,
@@ -72,6 +76,7 @@ RasterOverlayTileProvider::RasterOverlayTileProvider(
     : _pOwner(const_intrusive_cast<RasterOverlay>(pOwner)),
       _asyncSystem(asyncSystem),
       _pAssetAccessor(pAssetAccessor),
+      _pCreditSystem(pCreditSystem),
       _credit(credit),
       _pPrepareRendererResources(pPrepareRendererResources),
       _pLogger(pLogger),
@@ -80,7 +85,8 @@ RasterOverlayTileProvider::RasterOverlayTileProvider(
       _pPlaceholder(nullptr),
       _tileDataBytes(0),
       _totalTilesCurrentlyLoading(0),
-      _throttledTilesCurrentlyLoading(0) {}
+      _throttledTilesCurrentlyLoading(0),
+      _destructionCompleteDetails() {}
 
 RasterOverlayTileProvider::~RasterOverlayTileProvider() noexcept {
   // Explicitly release the placeholder first, because RasterOverlayTiles must
@@ -89,6 +95,23 @@ RasterOverlayTileProvider::~RasterOverlayTileProvider() noexcept {
     CESIUM_ASSERT(this->_pPlaceholder->getReferenceCount() == 1);
     this->_pPlaceholder = nullptr;
   }
+
+  if (this->_destructionCompleteDetails) {
+    this->_destructionCompleteDetails->promise.resolve();
+  }
+}
+
+CesiumAsync::SharedFuture<void>&
+RasterOverlayTileProvider::getAsyncDestructionCompleteEvent() {
+  if (!this->_destructionCompleteDetails) {
+    auto promise = this->_asyncSystem.createPromise<void>();
+    auto sharedFuture = promise.getFuture().share();
+    this->_destructionCompleteDetails.emplace(DestructionCompleteDetails{
+        std::move(promise),
+        std::move(sharedFuture)});
+  }
+
+  return this->_destructionCompleteDetails->future;
 }
 
 CesiumUtility::IntrusivePointer<RasterOverlayTile>