+++ CHANGES.md
@@ -4,8 +4,25 @@
 
 ##### Additions :tada:
 
-- Added `ImplicitTilingUtilities::getParentID` to derive the ID of the parent for a given tile ID.
 - Added `accessorView` to `PropertyAttributePropertyView` to retrieve the underlying `AccessorView`.
+
+### v0.50.0 - 2025-08-01
+
+##### Breaking Changes :mega:
+
+- The `RasterOverlayTileProvider` and `QuadtreeRasterOverlayTileProvider` constructors now require a `CreditSystem` parameter.
+
+##### Additions :tada:
+
+- Added `GeoJsonDocumentRasterOverlay` for displaying GeoJSON documents as a raster overlay.
+- Improved performance of `RasterizedPolygonsOverlay`, especially when using lots of cartographic polygons at once.
+- Added `ImplicitTilingUtilities::getParentID` to derive the ID of the parent for a given tile ID.
+- `IonRasterOverlay` now automatically handles refreshing the Cesium ion asset token as needed.
+- Added `CesiumIonAssetAccessor`, which is useful for implementing token refresh for Cesium ion assets.
+- Added `refreshTileProviderWithNewKey` method to `BingMapsRasterOverlay`.
+- Added `refreshTileProviderWithNewUrlAndHeaders` method to `TileMapServiceRasterOverlay`.
+- Added `getAsyncDestructionCompleteEvent` method to `RasterOverlayTileProvider`.
+- Added `getCreditSystem` method to `RasterOverlayTileProvider`.
 
 ### v0.49.0 - 2025-07-01